usleep is also used more than once
